apple-keyboard/host/requirements.txt
pynput==1.7.7
pyserial==3.5
simple-term-menu==1.6.4
